+++ CHANGELOG.md
@@ -28,11 +28,13 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1298 Unittest failed when on CPU2GPU case
 -   \#1359 Negative distance value returned when searching with HNSW index type
 -   \#1429 Server crashed when searching vectors using GPU
+-   \#1476 Fix vectors results bug when getting vectors from segments
 -   \#1484 Index type changed to IDMAP after compacted 
 -   \#1499 Fix duplicated ID number issue
 -   \#1491 Server crashed during adding vectors  
 -   \#1504 Avoid possible race condition between delete and search 
 -   \#1510 Add set interfaces for WAL configurations
+-   \#1511 Fix big integer cannot pass to server correctly
 
 ## Feature
 -   \#216 Add CLI to get server info